feature(mvc): move issues on board in a eficient way
why:
When many issues exists on a boardm then there exist performance
problems on the loading of them. To solve this paggination will be used.
The problem is that this would make stop working the way of how order
and move bettwen board_lits are handled. This try to prevent this problem.

How:
By send to the issue update endpoint with a ordinal_number, the order
(ordinal_number) of all the issues on the board_list are adaprted