fix: more_id was set to nil on relationship endpoint
The endpoint /board_lists/board_list_id/issues
has return a json with a links section. In this
link section, is the path for the next page.
The page use the more_id for paggination.
This more_id query parameter was null

Consecuences:
It was only posible to load first and second page
of the issues on the board_lists