feature: reenable registration feature test
**Context**: The registration feature test was disabled
when the cucumber code was refactored to run against
a real local server. To run against a local server,
the feature tests use the endpont "POST /api/v1/test_organizations".
This endpoint deletest and create a test organization. The problem
was that for the registation no organization with the name
used in the feature could exist, to not make it fail.

**Solution**: Create a endpoint to delete the test organization