feature: add support for global board
**Context**: Using global board lists, instead of project specific ones
should be posible. With global board list is meant, that it every project has the
same board lists and when they are shown on a project, a filter is applied
to his issues.

**Feature increment**: This add global board lists to new organization but disable them for
now. Everything should work as before without notice any diference. This is
one step required to totaly support global board lists.